CONFIG added, token validation added

+++ goals/main.py
@@ -1,5 +1,5 @@
 """Requests handlers."""
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import Depends, FastAPI, HTTPException, Request
 from fastapi.responses import JSONResponse
 from environ import to_config
 from prometheus_client import start_http_server, Counter
@@ -8,10 +8,11 @@ from sqlalchemy.orm import Session
 
 from goals.config import AppConfig
 from goals.database.crud import create_goal, get_user_goals, get_goal, \
-    get_all_metrics, delete_goal, update_goal
+    get_all_metrics, delete_goal, update_goal, correct_user_id
 from goals.database.models import Base
 from goals.database.initialization import get_database_url
 from goals.schemas import GoalBase, GoalUpdate
+from goals.util import get_credentials
 
 BASE_URI = "/goals"
 
@@ -34,42 +35,60 @@ def get_db() -> Session:
 
 
 @app.post(BASE_URI + "/{user_id}")
-async def add_goal_for_user(goal: GoalBase, user_id: str,
+async def add_goal_for_user(request: Request,
+                            goal: GoalBase, user_id: str,
                             session: Session = Depends(get_db)):
     """Create a new goal for user_id."""
+    creds = await get_credentials(request)
+    if creds["id"] != user_id:
+        raise HTTPException(status_code=403, detail="Invalid credentials")
     with session as open_session:
         _id = create_goal(session=open_session, goal=goal, user_id=user_id)
         return _id
 
 
 @app.get(BASE_URI + "/metrics")
-async def get_metrics(session: Session = Depends(get_db)):
+async def get_metrics(request: Request,
+                      session: Session = Depends(get_db)):
     """Return all metrics in database."""
+    creds = await get_credentials(request)
+    if not creds["role"] == "admin" and not creds["role"] == "user":
+        raise HTTPException(status_code=403, detail="Invalid credentials")
     with session as open_session:
         return get_all_metrics(open_session)
 
 
 @app.get(BASE_URI + "/{user_id}")
-async def get_goals(user_id: str,
+async def get_goals(request: Request, user_id: str,
                     session: Session = Depends(get_db)):
     """Return all goals in database."""
+    creds = await get_credentials(request)
+    if creds["id"] != user_id:
+        raise HTTPException(status_code=403, detail="Invalid credentials")
     with session as open_session:
         return get_user_goals(session=open_session, user_id=user_id)
 
 
 @app.delete(BASE_URI + "/{goal_id}")
-async def delete_user_goal(goal_id: int, session: Session = Depends(get_db)):
+async def delete_user_goal(request: Request,
+                           goal_id: int, session: Session = Depends(get_db)):
     """Delete goal with goal_id."""
+    creds = await get_credentials(request)
     if get_goal(session, goal_id) is None:
         raise HTTPException(status_code=404, detail="No such goal")
+    if correct_user_id(session, goal_id, creds["id"]) is False:
+        raise HTTPException(status_code=403, detail="Invalid credentials")
     with session as open_session:
         delete_goal(session=open_session, goal_id=goal_id)
 
 
 @app.patch(BASE_URI + "/{goal_id}")
-async def _update_goal(goal_update: GoalUpdate,
+async def _update_goal(request: Request, goal_update: GoalUpdate,
                        goal_id: int, session: Session = Depends(get_db)):
     """Update goal with goal_id."""
+    creds = await get_credentials(request)
+    if correct_user_id(session, goal_id, creds["id"]) is False:
+        raise HTTPException(status_code=403, detail="Invalid credentials")
     with session as open_session:
         if get_goal(open_session, goal_id=goal_id) is None:
             raise HTTPException(status_code=404, detail="No such goal")
